유저정보 별도 컴포넌트화 및 라우터에 개별유저 뷰 지정
