+++ src/components/common/MainToolbar.vue
@@ -6,9 +6,9 @@
   >
     <ul>
       <li v-bind:class="{ active: menu == 'Main'}"><router-link to="/">추천</router-link></li>
-      <li v-bind:class="{ active: menu == 'Feeds'}"><router-link to="feeds">피드</router-link></li>
-      <li v-bind:class="{ active: menu == 'Activities'}"><router-link to="activities">활동</router-link></li>
-      <li v-bind:class="{ active: menu == 'My'}"><router-link to="my">MY</router-link></li>
+      <li v-bind:class="{ active: menu == 'Feeds'}"><router-link to="/feeds">피드</router-link></li>
+      <li v-bind:class="{ active: menu == 'Activities'}"><router-link to="/activities">활동</router-link></li>
+      <li v-bind:class="{ active: menu == 'My'}"><router-link to="/my">MY</router-link></li>
     </ul>
   </v-toolbar>
 </template>
