+++ src/router/index.js
@@ -11,6 +11,7 @@ import Lease from '@/pages/center/Lease'
 import Delegate from '@/pages/center/Delegate'
 import Transactions from '@/pages/center/Transactions'
 import RequestVote from '@/pages/center/RequestVote'
+import UserProfile from '@/components/UserProfile'
 
 import Create from '@/pages/Create'
 
@@ -50,6 +51,11 @@ const router = new Router({
       component: My
     },
     {
+      path: '/user/:username',
+      name: 'User',
+      component: UserProfile
+    },
+    {
       path: '/center/lease',
       name: 'Lease',
       component: Lease
@@ -77,7 +83,7 @@ const router = new Router({
   ]
 })
 router.beforeResolve((to, from, next) => {
-  const defaultLayout = ['Feeds', 'Activities', 'My']
+  const defaultLayout = ['Feeds', 'Activities', 'My', 'User']
   const centerLayout = ['Lease', 'Delegate', 'RequestVote', 'Transactions']
   // const appLayout = ['Delegate']
   if (to.name === 'Main') {
